+++ src/structs/Interval.jl
@@ -1,0 +1,4 @@
+struct Interval{T}
+    from::T
+    to::T
+end