+++ src/structs/LinearScale.jl
@@ -1,0 +1,5 @@
+struct LinearScale{T}
+    input::Interval{T}
+    output::Interval{T}
+    clamp::Bool
+end