+++ src/structs/objects/CoordinateSystem.jl
@@ -1,0 +1,13 @@
+struct CoordinateSystem
+    left::Point
+    right::Point
+    bottom::Point
+    top::Point
+    fct::Function
+    mainwidth::Float64
+    gridwidth::Float64
+    step_size_x::Float64
+    step_size_y::Float64
+    maincolor::Any
+    gridcolor::Any
+end